When reshape, working data are reshaped. N change.
An external forced reshaped is introduced.
The number of nucleotide is shrunken to [30, 60] from [4, 110]
Furthermore, when a reshaped is actuated, the working data is the one reshaped and not (as before) the original one with just the rupture force better calculated
Moreover, the formula that computes the inverse of the WLC model is updated. Now the extreme case D = 0 is covered. The notation adopted is the same on Wikipedia.
Analysis are then reperformed